[Refactor] 리팩터링데이 - 버그 6개 수정

+++ src/main/java/com/tilguys/matilda/common/auth/strategy/AccessJwtTokenCookieCreateStrategy.java
@@ -17,7 +17,7 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class AccessJwtTokenCookieCreateStrategy implements JwtCookieCreateStrategy {
 
-    private static final long ACCESS_TOKEN_EXPIRE_TIME = 1000 * 60 * 60; // 1시간
+    private static final long ACCESS_TOKEN_EXPIRE_TIME = (long) 1000 * 60 * 60; // 1시간
     private final Key key;
 
     @Override
